+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Dates are formatted in YYMMDD format. Each major version has a codename, this is noted with brackets beside the version number.
 
+## [0.0.9] - Jan 19, 2023
+
+### Added
+
+- Build calendar screen
+- Fetch calendar events from Google Calendar
+- Refresh calendar events when user swipes down and holds
+- Refresh posts when user swipes down and holds
+
+### Fixed
+
+- Unicode characters showing up in post title text
+
 ## [0.0.8] - Jan 14, 2023
 
 ### Added
@@ -14,7 +27,7 @@ Dates are formatted in YYMMDD format. Each major version has a codename, this is
 
 ### Fixed
 
-- Unicode characters showing up in title text
+- Unicode characters showing up in home screen title text
 
 ## [0.0.7] - Jan 5, 2023